# api-docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProductCreateRequest": {
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "price": {
            "title": "Price",
            "type": "number"
          },
          "sku": {
            "title": "Sku",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "variant_options": {
            "items": {
              "type": "object"
            },
            "title": "Variant Options",
            "type": "array"
          }
        },
        "required": [
          "name",
          "sku",
          "price"
        ],
        "title": "ProductCreateRequest",
        "type": "object"
      },
      "ProductListResponse": {
        "properties": {
          "products": {
            "items": {
              "$ref": "#/components/schemas/ProductResponse"
            },
            "title": "Products",
            "type": "array"
          }
        },
        "required": [
          "products"
        ],
        "title": "ProductListResponse",
        "type": "object"
      },
      "ProductResponse": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "landing_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Landing Id"
          },
          "landing_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Landing Slug"
          },
          "landing_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Landing Status"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "price": {
            "title": "Price",
            "type": "number"
          },
          "retired_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retired At"
          },
          "sku": {
            "title": "Sku",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "variant_options": {
            "items": {
              "type": "object"
            },
            "title": "Variant Options",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "sku",
          "price",
          "description",
          "status"
        ],
        "title": "ProductResponse",
        "type": "object"
      },
      "ProductUpdateRequest": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "sku": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sku"
          }
        },
        "title": "ProductUpdateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "The same session token supplied as `Authorization: Bearer <token>`, for clients that cannot hold cookies. Accepted everywhere the cookie is.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "Session cookie issued by `POST /api/auth/login`. `HttpOnly`, `SameSite=Strict`, and `Secure` outside local development, so a browser sends it automatically and page scripts cannot read it. This is the transport browser clients should use.",
        "in": "cookie",
        "name": "session",
        "type": "apiKey"
      }
    }
  },
  "info": {
...
    "summary": "Cash-on-delivery commerce API for a single merchant: public landing delivery and checkout, plus the authenticated Administrator surface.",
    "title": "Easy Funnel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/products": {
      "get": {
        "description": "List products, excluding retired by default.",
        "operationId": "list_products_api_admin_products_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_retired",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Retired",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required."
                },
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No valid Administrator session. Returned identically for a missing, expired, tampered or otherwise invalid token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "List Products",
        "tags": [
          "products"
        ]
      },
      "post": {
        "description": "Create a new product with an atomic draft landing.",
        "operationId": "create_product_api_admin_products_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required."
                },
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No valid Administrator session. Returned identically for a missing, expired, tampered or otherwise invalid token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create Product",
        "tags": [
          "products"
        ]
      }
    },
    "/api/admin/products/{product_id}": {
      "delete": {
        "description": "Soft-delete (retire) a product.",
        "operationId": "retire_product_api_admin_products__product_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required."
                },
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No valid Administrator session. Returned identically for a missing, expired, tampered or otherwise invalid token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Retire Product",
        "tags": [
          "products"
        ]
      },
      "get": {
        "description": "Get a specific product.",
        "operationId": "get_product_api_admin_products__product_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required."
                },
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No valid Administrator session. Returned identically for a missing, expired, tampered or otherwise invalid token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get Product",
        "tags": [
          "products"
        ]
      },
      "patch": {
        "description": "Validate and update supplied fields on a non-retired product.",
        "operationId": "update_product_api_admin_products__product_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required."
                },
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No valid Administrator session. Returned identically for a missing, expired, tampered or otherwise invalid token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update Product",
        "tags": [
          "products"
        ]
      }
    },
    "/api/admin/products/{product_id}/activate": {
      "post": {
        "description": "Activate a paused product.",
        "operationId": "activate_product_api_admin_products__product_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required."
                },
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No valid Administrator session. Returned identically for a missing, expired, tampered or otherwise invalid token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Activate Product",
        "tags": [
          "products"
        ]
      }
    },
...
    "/api/admin/products/{product_id}/pause": {
      "post": {
        "description": "Pause an active product.",
        "operationId": "pause_product_api_admin_products__product_id__pause_post",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required."
                },
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "No valid Administrator session. Returned identically for a missing, expired, tampered or otherwise invalid token."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Pause Product",
        "tags": [
          "products"
        ]
      }
    },
...
  },
...
}